Add BetterMersenneTwisterCracker

+++ CasinoRoyale/Models/Result.cs
@@ -3,7 +3,7 @@
     public class Result
     {
         public Account Account { get; set; }
-        public int RealNumber { get; set; }
+        public long RealNumber { get; set; }
         public bool IsVictory { get; set; }
         public string Message { get; set; }
     }
